pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "factory-commentary"
version = "0.1.0"
description = "A multi-channel YouTube automation system that converts SRT files into CapCut video drafts"
readme = "README.md"
requires-python = ">=3.9"
authors = [
    {name = "YouTube Automation Team"}
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    # Core dependencies
    "pandas>=1.5.0",
    "requests>=2.28.0",
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0",

    # LLM and AI
    "google-generativeai>=0.4.0",
    "openai>=1.0.0",

    # Image processing
    "pillow>=10.0.0",
    "numpy>=1.24.0",

    # CLI and utilities
    "click>=8.0.0",
    "tqdm>=4.65.0",

    # UI components (if needed)
    "streamlit>=1.28.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
    "mypy>=1.0.0",
]

[project.urls]
Homepage = "https://github.com/your-username/factory_commentary"
Repository = "https://github.com/your-username/factory_commentary"

[project.scripts]
factory-commentary = "commentary_02_srt2images_timeline.tools.factory:main"
factory-ch02 = "commentary_02_srt2images_timeline.tools.factory:main"  # LEGACY alias - use factory-commentary instead

# Use find_packages with explicit configuration to handle the multiple project structure
[tool.setuptools.packages.find]
# Look in the project root for packages
where = ["."]
# Only include the top-level packages we want to be part of the distribution
include = ["factory_common*", "commentary_02_srt2images_timeline*", "audio_tts_v2*", "script_pipeline*"]
# Don't search in these directories
exclude = ["tests*", "docs*", "logs*", "output*", "data*", "asset*", "thumbnails*", "remotion*"]